Update LocalFS set up script to take a list
The object name and directory path were hardcoded in the script.  We
need to be able to pass in that information from the command line.

Running the following command

```
./bin/instance run scripts/set_up_localfs_for_pdfs.py pdfs:/var/www/files pdfs2:/var/www/files2
```

will create 2 "LocalFS" objects in the plone site called "pdfs" and
"pdfs2", pointing to /var/www/files and /var/www/files2 in the
filesystem respectively.