Make all portal roles assignable in scripts/addRhaptosUser.zctl
For some reason, not all portal roles are assignable on the "Users and
Groups Administration" page.  When looking at all the users, some role
checkboxes are disabled, making it impossible to assign to the user
using the admin interface.

`scripts/addRhaptosUser.zctl` is what we use to create users on the site
so just add some code to make all the portal roles assignable.